test: move AirtableClientTest to Unit folder
