--- tests/Functionnal/Mock/Customer.php
@@ -1,10 +1,0 @@
-<?php
-
-namespace Yoanbernabeu\AirtableClientBundle\Tests\Functionnal\Mock;
-
-class Customer
-{
-    public $id;
-    public $firstName;
-    public $lastName;
-}
