+++ tests/Unit/AirtableRecordTest.php
@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 namespace Yoanbernabeu\AirtableClientBundle\Tests\Unit;
 
 use PHPUnit\Framework\TestCase;
